corrected demo

+++ src/components/Balance/Balance.jsx
@@ -1,12 +1,19 @@
 import { useState, useEffect } from "react";
 import { toast } from "react-toastify";
 import { useBalance } from "../../context/BalanceContext";
+import { useDemo } from "../../context/DemoContext"; ////nowe
 import "./Balance.css";
 import { BalanceModal } from "../BalanceModal/BalanceModal";
 
 const Balance = () => {
 	const { balance, loading, updateBalance, calculateTransactionTotal } =
 		useBalance();
+
+	const { setDemoBalance, demoBalance } = useDemo(); ////nowe
+
+	const user = JSON.parse(localStorage.getItem("user")); ////nowe
+	const isDemo = user?.email === "guest@demo.com"; ////nowe
+
 	const [input, setInput] = useState(balance || "0.00");
 	const [isEditing, setIsEditing] = useState(false);
 	const [showModal, setShowModal] = useState(false);
@@ -14,16 +21,13 @@ const Balance = () => {
 
 	useEffect(() => {
 		const balanceConfirmed = localStorage.getItem("balanceConfirmed");
-		if (
-			!loading &&
-			parseFloat(balance) === 0 &&
-			balanceConfirmed !== "true"
-		) {
+		const val = isDemo ? demoBalance : balance; ////nowe
+		if (!loading && parseFloat(val) === 0 && balanceConfirmed !== "true") {
 			setShowModal(true);
 		} else {
 			setShowModal(false);
 		}
-	}, [balance, loading]);
+	}, [balance, demoBalance, loading]); ////zmieniono zależności
 
 	const handleChange = (e) => {
 		const inputValue = e.target.value.replace(" EUR", "");
@@ -40,7 +44,7 @@ const Balance = () => {
 	const handleBlur = () => {
 		if (!isEditing) return;
 		if (!input || input === "") {
-			setInput(balance);
+			setInput(isDemo ? demoBalance : balance); ////nowe
 		}
 		setIsEditing(false);
 	};
@@ -58,13 +62,20 @@ const Balance = () => {
 		}
 
 		try {
+			if (isDemo) {
+				setDemoBalance(newInitialBalance); ////nowe
+				setInitialBalance(newInitialBalance); ////nowe
+				localStorage.setItem("balanceConfirmed", "true"); ////nowe
+				setShowModal(false); ////nowe
+				document.activeElement.blur(); ////nowe
+				return; ////nowe
+			}
+
 			const totalTransactions = await calculateTransactionTotal();
 			const updatedBalance = newInitialBalance + totalTransactions;
 
 			await updateBalance(updatedBalance);
-
 			setInitialBalance(newInitialBalance);
-
 			localStorage.setItem("balanceConfirmed", "true");
 			setShowModal(false);
 			document.activeElement.blur();
@@ -92,9 +103,9 @@ const Balance = () => {
 								value={
 									isEditing
 										? input
-										: `${parseFloat(balance).toFixed(
-												2
-										  )} EUR`
+										: `${parseFloat(
+												isDemo ? demoBalance : balance ////nowe
+										  ).toFixed(2)} EUR`
 								}
 								onChange={handleChange}
 								onFocus={handleFocus}
